refactor: AddOrganizationController tests
