+++ tests/presentation/mocks/mock-organization.ts
@@ -1,4 +1,4 @@
-import { CheckOrganizationById } from '../../../src/domain/usecases'
+import { AddOrganization, CheckOrganizationById } from '../../../src/domain/usecases'
 
 export class CheckOrganizationByIdSpy implements CheckOrganizationById {
   id: string
@@ -7,5 +7,13 @@ export class CheckOrganizationByIdSpy implements CheckOrganizationById {
   async checkById (id: string): Promise<boolean> {
     this.id = id
     return this.result
+  }
+}
+
+export class AddOrganizationSpy implements AddOrganization {
+  params: AddOrganization.Params
+
+  async add (organization: AddOrganization.Params): Promise<void> {
+    this.params = organization
   }
 }
